Modificacion prueba práctica
Se modificó el proyecto Protocolo con lo requerido y se actualizó el código.
