+++ Cliente/Form1.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Linq;
-using System.Text;
 using System.Windows.Forms;
 using System.Net.Sockets;
 using Protocolo;
@@ -12,15 +10,17 @@ namespace Cliente
         private TcpClient remoto;
         private NetworkStream flujo;
 
+        //Se inicializan los componentes del form
         public FrmValidador()
         {
             InitializeComponent();
         }
-
+        //Al cargar el form se establece la conexión con el server
         private void FrmValidador_Load(object sender, EventArgs e)
         {
             try
             {
+                //Se usa un try en caso de que al conectarse se genera una excepcion
                 remoto = new TcpClient("127.0.0.1", 8080);
                 flujo = remoto.GetStream();
             }
@@ -29,12 +29,9 @@ namespace Cliente
                 MessageBox.Show("No se puedo establecer conexión " + ex.Message,
                     "ERROR");
             }
-            finally 
-            {
-                flujo?.Close();
-                remoto?.Close();
-            }
+            //Se elimina los finally porque generan excepciones al momento de la carga.
 
+            //Se deshabilita ciertos elementos para un mejor manejo del programa
             panPlaca.Enabled = false;
             chkLunes.Enabled = false;
             chkMartes.Enabled = false;
@@ -45,6 +42,7 @@ namespace Cliente
             chkSabado.Enabled = false;
         }
 
+        //Al hacer click en el boton iniciar se hace la validacion de los datos ingresados
         private void btnIniciar_Click(object sender, EventArgs e)
         {
             string usuario = txtUsuario.Text;
@@ -62,13 +60,15 @@ namespace Cliente
                 Parametros = new[] { usuario, contraseña }
             };
             
-            Respuesta respuesta = HazOperacion(pedido);
+            //Se hace uso de la operación con el Protoclo definido en la bibliotecca de clases
+            Respuesta respuesta = Protocolo.Protocolo.HazOperacion(pedido, flujo);
             if (respuesta == null)
             {
                 MessageBox.Show("Hubo un error", "ERROR");
                 return;
             }
 
+            //Si por parte del server se recibe un OK el acceso es permitido y se puede proseguir, caso contrario se nos deniega el acceso
             if (respuesta.Estado == "OK" && respuesta.Mensaje == "ACCESO_CONCEDIDO")
             {
                 panPlaca.Enabled = true;
@@ -86,66 +86,30 @@ namespace Cliente
             }
         }
 
-        private Respuesta HazOperacion(Pedido pedido)
-        {
-            if(flujo == null)
-            {
-                MessageBox.Show("No hay conexión", "ERROR");
-                return null;
-            }
-            try
-            {
-                byte[] bufferTx = Encoding.UTF8.GetBytes(
-                    pedido.Comando + " " + string.Join(" ", pedido.Parametros));
-                
-                flujo.Write(bufferTx, 0, bufferTx.Length);
-
-                byte[] bufferRx = new byte[1024];
-                
-                int bytesRx = flujo.Read(bufferRx, 0, bufferRx.Length);
-                
-                string mensaje = Encoding.UTF8.GetString(bufferRx, 0, bytesRx);
-                
-                var partes = mensaje.Split(' ');
-                
-                return new Respuesta
-                {
-                    Estado = partes[0],
-                    Mensaje = string.Join(" ", partes.Skip(1).ToArray())
-                };
-            }
-            catch (SocketException ex)
-            {
-                MessageBox.Show("Error al intentar transmitir " + ex.Message,
-                    "ERROR");
-            }
-            finally 
-            {
-                flujo?.Close();
-                remoto?.Close();
-            }
-            return null;
-        }
-
+       
         private void btnConsultar_Click(object sender, EventArgs e)
         {
+            //Se almacenan los datos ingresados por el cliente
             string modelo = txtModelo.Text;
             string marca = txtMarca.Text;
             string placa = txtPlaca.Text;
             
+            // Se instancia un pedido
             Pedido pedido = new Pedido
             {
                 Comando = "CALCULO",
                 Parametros = new[] { modelo, marca, placa }
             };
             
-            Respuesta respuesta = HazOperacion(pedido);
+            // Se hace uso del Protocolo para realizar la operación 
+            Respuesta respuesta = Protocolo.Protocolo.HazOperacion(pedido, flujo);
             if (respuesta == null)
             {
                 MessageBox.Show("Hubo un error", "ERROR");
                 return;
             }
 
+            //En caso de que la respuesta sea negativa se presenta un mensaje de error en la solicitud
             if (respuesta.Estado == "NOK")
             {
                 MessageBox.Show("Error en la solicitud.", "ERROR");
@@ -157,6 +121,7 @@ namespace Cliente
             }
             else
             {
+                // Si es aceptado, se marca el checkbox edl día en el que corresponde
                 var partes = respuesta.Mensaje.Split(' ');
                 MessageBox.Show("Se recibió: " + respuesta.Mensaje,
                     "INFORMACIÓN");
@@ -209,23 +174,27 @@ namespace Cliente
             }
         }
 
+        //Se hace el llamado al numero de consultas que ha hecho un usuario
         private void btnNumConsultas_Click(object sender, EventArgs e)
         {
             String mensaje = "hola";
             
+            //Se instancia un pedido
             Pedido pedido = new Pedido
             {
                 Comando = "CONTADOR",
                 Parametros = new[] { mensaje }
             };
 
-            Respuesta respuesta = HazOperacion(pedido);
+            //Se evalua y realiza la operación con el Protocolo utilizado
+            Respuesta respuesta = Protocolo.Protocolo.HazOperacion(pedido, flujo);
             if (respuesta == null)
             {
                 MessageBox.Show("Hubo un error", "ERROR");
                 return;
             }
-
+            
+            //Se evalua la respuesta en caso de ser negativa se muestra el mensaje de error, sino se devuelve el numero de solicitudes realizadas.
             if (respuesta.Estado == "NOK")
             {
                 MessageBox.Show("Error en la solicitud.", "ERROR");
@@ -239,6 +208,7 @@ namespace Cliente
             }
         }
 
+        //Se debe cerrar las conexiones con el servidor en caso de que se cierre el formulario, caso contrario la conexion queda abierta
         private void FrmValidador_FormClosing(object sender, FormClosingEventArgs e)
         {
             if (flujo != null)
