+++ Servidor/Servidor.cs
@@ -1,7 +1,36 @@
-﻿using System;
+﻿// ************************************************************************
+// Practica 07
+// Esteban Chicango
+// Fecha de realización: 22/06/2024
+// Fecha de entrega: 26/06/2024
+//
+// Resultados:
+// Como se requería, se implmentó la claseProtocolo donde se maneja las operaciones y acciones entre el cliente y el servidor
+// de la misma forma, estas crean una mejor comunicación y hacen más entendible el protocolo utilizado.
+// Se cambió el mensaje que decía que se oescucha en el puerto 5000 cuando en realidad lo hacemos en el  8080
+// El error sucedido encontrado en clase fue eliminado, se quitó los finally en el cliente al cerrar las conexiones.
+// Este error aparece pues cierran la conexion cuando el form se está cargando, entonces salta la excpecion.
+//
+//Conclusiones:
+// - El utilizar una clase protocolo facilita enormemente le utilización y entendimiento del código para pedidos y 
+// respuestas, alivianando tambien el volumen de codigo en el servidor y en el cliente.
+// - Las validaciones en la mayoría de los casos ayuda al cliente a efectuar de manera correcta el programa sin que
+// este se caiga, de no utilizarlas el cliente puede generar complicaciones y el programa puede lanzar excepciones.
+//
+//Recomendaciones:
+// -Cerrar las conexiones del cliente con el servidor al momento que un formulario se cierre, de no hacer esto, la conexión
+// segruirá activa y puede generar complicaciones.
+// - Utilizar puntos de para con el fin de analizar paso a paso los eventos que van sucediendo en el codigo, esto ayuda a
+// detectar posibles fallos y entender de mejor manera los procesos realizados.
+// - Utilizar un protocolo bien definido para las peticiones y respuetas, esto facilita la comunicación y entendimiento entre 
+// ambas partes, el cliente y el servidor.
+// *****************************************************************************
+
+
+
+using System;
 using System.Collections.Generic;
 using System.Text;
-using System.Text.RegularExpressions;
 using System.Threading;
 using System.Net;
 using System.Net.Sockets;
@@ -11,6 +40,7 @@ namespace Servidor
 {
     class Servidor
     {
+        // Se instancian los elementos necesarios para la conexión
         private static TcpListener escuchador;
         private static Dictionary<string, int> listadoClientes
             = new Dictionary<string, int>();
@@ -19,10 +49,13 @@ namespace Servidor
         {
             try
             {
+                //Se define un escucha que trabajará en el puerto 8080
                 escuchador = new TcpListener(IPAddress.Any, 8080);
                 escuchador.Start();
-                Console.WriteLine("Servidor inició en el puerto 5000...");
+                //Se modifica el mensaje para que indique el puerto donde realmente está escuchando 8080
+                Console.WriteLine("Servidor inició en el puerto 8080...");
 
+                //Se usa el lazo while para manejar distintos clientes mediante hilos
                 while (true)
                 {
                     TcpClient cliente = escuchador.AcceptTcpClient();
@@ -33,6 +66,7 @@ namespace Servidor
             }
             catch (SocketException ex)
             {
+                //Se controla el error y se muestra un mensaje si es que no se puede establecer la conexion
                 Console.WriteLine("Error de socket al iniciar el servidor: " +
                     ex.Message);
             }
@@ -42,8 +76,10 @@ namespace Servidor
             }
         }
 
+        //Se maneja la comunicación con el cliente
         private static void ManipuladorCliente(object obj)
         {
+            //Se intancia un cliente TCP y un flujo con el que se va a trabajar
             TcpClient cliente = (TcpClient)obj;
             NetworkStream flujo = null;
             try
@@ -53,16 +89,20 @@ namespace Servidor
                 byte[] bufferRx = new byte[1024];
                 int bytesRx;
 
+                //Se lee y procesa los mensajes del cliente
                 while ((bytesRx = flujo.Read(bufferRx, 0, bufferRx.Length)) > 0)
                 {
+                    //Se procesa el mensaje recibido por parte del cliente (peticion), se lo transforma y se muestra el mensaje en consola del server
                     string mensajeRx =
                         Encoding.UTF8.GetString(bufferRx, 0, bytesRx);
                     Pedido pedido = Pedido.Procesar(mensajeRx);
-                    Console.WriteLine("Se recibio: " + pedido);
+                    Console.WriteLine("Se recibió: " + pedido);
 
                     string direccionCliente =
                         cliente.Client.RemoteEndPoint.ToString();
-                    Respuesta respuesta = ResolverPedido(pedido, direccionCliente);
+                    //Se hace uso del protocolo creado
+                    Respuesta respuesta = Protocolo.Protocolo.ResolverPedido(pedido, direccionCliente, listadoClientes);
+                    //En consola del server se muestra lo que se envió al cliente
                     Console.WriteLine("Se envió: " + respuesta);
 
                     bufferTx = Encoding.UTF8.GetBytes(respuesta.ToString());
@@ -70,6 +110,7 @@ namespace Servidor
                 }
 
             }
+            //Se maneja la excepción al manejar al cliente
             catch (SocketException ex)
             {
                 Console.WriteLine("Error de socket al manejar el cliente: " + ex.Message);
@@ -81,111 +122,6 @@ namespace Servidor
             }
         }
 
-        private static Respuesta ResolverPedido(Pedido pedido, string direccionCliente)
-        {
-            Respuesta respuesta = new Respuesta
-            { Estado = "NOK", Mensaje = "Comando no reconocido" };
-
-            switch (pedido.Comando)
-            {
-                case "INGRESO":
-                    if (pedido.Parametros.Length == 2 &&
-                        pedido.Parametros[0] == "root" &&
-                        pedido.Parametros[1] == "admin20")
-                    {
-                        respuesta = new Random().Next(2) == 0
-                            ? new Respuesta 
-                            { Estado = "OK", 
-                                Mensaje = "ACCESO_CONCEDIDO" }
-                            : new Respuesta 
-                            { Estado = "NOK", 
-                                Mensaje = "ACCESO_NEGADO" };
-                    }
-                    else
-                    {
-                        respuesta.Mensaje = "ACCESO_NEGADO";
-                    }
-                    break;
-
-                case "CALCULO":
-                    if (pedido.Parametros.Length == 3)
-                    {
-                        string modelo = pedido.Parametros[0];
-                        string marca = pedido.Parametros[1];
-                        string placa = pedido.Parametros[2];
-                        if (ValidarPlaca(placa))
-                        {
-                            byte indicadorDia = ObtenerIndicadorDia(placa);
-                            respuesta = new Respuesta
-                            { Estado = "OK", 
-                                Mensaje = $"{placa} {indicadorDia}" };
-                            ContadorCliente(direccionCliente);
-                        }
-                        else
-                        {
-                            respuesta.Mensaje = "Placa no válida";
-                        }
-                    }
-                    break;
-
-                case "CONTADOR":
-                    if (listadoClientes.ContainsKey(direccionCliente))
-                    {
-                        respuesta = new Respuesta
-                        { Estado = "OK",
-                            Mensaje = listadoClientes[direccionCliente].ToString() };
-                    }
-                    else
-                    {
-                        respuesta.Mensaje = "No hay solicitudes previas";
-                    }
-                    break;
-            }
-
-            return respuesta;
-        }
-
-        private static bool ValidarPlaca(string placa)
-        {
-            return Regex.IsMatch(placa, @"^[A-Z]{3}[0-9]{4}$");
-        }
-
-        private static byte ObtenerIndicadorDia(string placa)
-        {
-            int ultimoDigito = int.Parse(placa.Substring(6, 1));
-            switch (ultimoDigito)
-            {
-                case 1: 
-                case 2: 
-                    return 0b00100000; // Lunes
-                case 3: 
-                case 4: 
-                    return 0b00010000; // Martes
-                case 5: 
-                case 6: 
-                    return 0b00001000; // Miércoles
-                case 7: 
-                case 8: 
-                    return 0b00000100; // Jueves
-                case 9: 
-                case 0: 
-                    return 0b00000010; // Viernes
-                default: 
-                    return 0;
-            }
-        }
-
-        private static void ContadorCliente(string direccionCliente)
-        {
-            if (listadoClientes.ContainsKey(direccionCliente))
-            {
-                listadoClientes[direccionCliente]++;
-            }
-            else
-            {
-                listadoClientes[direccionCliente] = 1;
-            }
-        }
 
     }
 }
